feat:全量头部注释优化

+++ 数学/多项式/2-数列与生成函数/BostanMori·常系数线性递推.cpp
@@ -1,26 +1,31 @@
+#include "aizalib.h"
 #include "0-base/Poly·多项式全家桶.hpp"
+/*
+ * Bostan-Mori Algorithm (常系数线性递推与有理分式展开)
+ *
+ * Overview:
+ *      求解有理分式系数 [x^n] P(x)/Q(x) 以及 d 阶常系数线性递推的第 n 项。
+ *      利用倍增转换 P(x)/Q(x) = (P(x)*Q(-x)) / (Q(x)*Q(-x))，分母变为关于 x^2
+ *      的偶多项式，根据 n 的奇偶性提取分子的奇偶项，每次将求解规模 n 折半，在 O(d
+ *      log d log n) 时间内求出远项 a_n，大幅优于传统多项式取模。
+ *
+ * API:
+ *     bostan_mori(n, P, Q)       — 求解 [x^n] P(x)/Q(x)，复杂度 O(d log d log n)。
+ *     linear_recurrence(n, c, a) — 求解满足 a_n = sum_{i=1..d} c_i * a_{n-i} 的第
+ *                                   n 项，其中 c = {c1, ..., cd}, a = {a0, ...,
+ *                                   a_{d-1}}。复杂度 O(d log d log n)。
+ *
+ * Notes:
+ *      1. 要求 Q[0] != 0。
+ *      2. n 为 64 位整数 (i64)，下标 0-based。
+ *
+ * Related:
+ *      数学/多项式/2-数列与生成函数/BerlekampMassey·最短线性递推.cpp:
+ *      求解最短递推式。
+ */
 
 namespace poly_ext {
 
-/**
- * Bostan-Mori(常系数线性递推)
- * 算法介绍: 用于求解有理分式 [x^n] P(x)/Q(x) 和常系数线性递推数列第 n 项。数学原理见文字部分。
- * 模板参数: 无
- * Interface: 
- *  - bostan_mori(n, P, Q)       — 求 [x^n] P(x)/Q(x)
- *  - linear_recurrence(n, c, a) — 求常系数线性递推数列第 n 项
- * Note:
- *      1. Time: O(d log d log n), d = deg(Q)
- *      2. Space: O(d)
- */
-
-/**
- * @brief Bostan-Mori 算法求 [x^n] P(x)/Q(x)
- * @param n 求第 n 项的系数 (0-indexed)
- * @param P 分子多项式
- * @param Q 分母多项式 (需要保证 Q[0] != 0)
- * @return [x^n] P(x)/Q(x) 取模后的值
- */
 int bostan_mori(i64 n, Poly P, Poly Q) {
     while (n > 0) {
         auto Q_neg = Q;
@@ -43,15 +48,9 @@ int bostan_mori(i64 n, Poly P, Poly Q) {
     return mul(P.empty() ? 0 : P[0], inv(Q[0]));
 }
 
-/**
- * @brief 求线性递推数列第 n 项
- * 满足 a_n = \sum_{i=1}^d c_i a_{n-i}
- * @param n 要求第 n 项 (0-indexed)
- * @param c 递推系数 {c_1, c_2, ..., c_d}
- * @param a 初值 {a_0, a_1, ..., a_{d-1}}
- * @return a_n 取模后的值
- */
-int linear_recurrence(long long n, const std::vector<int>& c, const std::vector<int>& a) {
+int linear_recurrence(
+    i64 n, const std::vector<int>& c, const std::vector<int>& a
+) {
     int d = c.size();
     if (n < d) return a[n];
